refactor(#8): spi의 R2DBC 사용 인터페이스를 persistence로 위치 이동
