+++ src/main/java/com/aetheri/infrastructure/persistence/repository/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Service Provider Interface (SPI) 이며 이 패키지는 R2DBC를 사용하기 위한 인터페이스들을 모아둔 패키지이다.
  * */
-package com.aetheri.domain.adapter.out.r2dbc.spi;
+package com.aetheri.infrastructure.persistence.repository;
